@@ -8,29 +8,29 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.7.1",
-    "axios": "^0.24.0",
-    "classnames": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1"
+    "@reduxjs/toolkit": "^1.9.3",
+    "axios": "^1.3.4",
+    "classnames": "^2.3.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.2"
   },
   "devDependencies": {
-    "@types/node": "^17.0.5",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-redux": "^7.1.21",
-    "@typescript-eslint/eslint-plugin": "^5.8.1",
-    "@typescript-eslint/parser": "^5.8.1",
-    "@vitejs/plugin-react": "^1.0.7",
-    "autoprefixer": "^10.4.1",
-    "eslint": "^8.5.0",
-    "eslint-plugin-react": "^7.28.0",
-    "postcss": "^8.4.5",
-    "sass": "^1.45.1",
-    "tailwindcss": "^3.0.11",
-    "typescript": "^4.4.4",
-    "vite": "^2.7.2"
+    "@types/node": "^18.14.4",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-redux": "^7.1.25",
+    "@typescript-eslint/eslint-plugin": "^5.54.0",
+    "@typescript-eslint/parser": "^5.54.0",
+    "@vitejs/plugin-react": "^3.1.0",
+    "autoprefixer": "^10.4.13",
+    "eslint": "^8.35.0",
+    "eslint-plugin-react": "^7.32.2",
+    "postcss": "^8.4.21",
+    "sass": "^1.58.3",
+    "tailwindcss": "^3.2.7",
+    "typescript": "^4.9.5",
+    "vite": "^4.1.4"
   }
 }
